[tools] Extract OptimizeGeneratedCode into its own class.
Extract the static optimization methods from OptimizeGeneratedCodeHandler into
a new standalone OptimizeGeneratedCode class. The handler class now only contains
the linker-specific scaffolding and delegates to the new class.

This makes the optimization logic reusable without depending on the linker's
ExceptionalMarkHandler base class.
