+++ tools/dotnet-linker/BackingFieldDelayHandler.cs
@@ -84,9 +84,9 @@ namespace Xamarin.Linker {
 							var store_field = ins;
 							var load_null = ins.Previous;
 							var load_this = ins.Previous.Previous;
-							if (OptimizeGeneratedCodeHandler.ValidateInstruction (app, method, store_field, operation, Code.Stfld) &&
-								OptimizeGeneratedCodeHandler.ValidateInstruction (app, method, load_null, operation, Code.Ldnull) &&
-								OptimizeGeneratedCodeHandler.ValidateInstruction (app, method, load_this, operation, Code.Ldarg_0)) {
+							if (OptimizeGeneratedCode.ValidateInstruction (app, method, store_field, operation, Code.Stfld) &&
+								OptimizeGeneratedCode.ValidateInstruction (app, method, load_null, operation, Code.Ldnull) &&
+								OptimizeGeneratedCode.ValidateInstruction (app, method, load_this, operation, Code.Ldarg_0)) {
 								store_field.OpCode = OpCodes.Nop;
 								load_null.OpCode = OpCodes.Nop;
 								load_this.OpCode = OpCodes.Nop;
